cluster: Improve Custom Resource Definition handling
Rather than using Unstructured, use the correct CRD packages out of apiserver extensions.
Move all commands to use that rather than the dynamic client.

Also, split out the utils function and integrate apiserver in them.
